# [IMP] purchase: speed up dashboard rfq sent computation.