--- a/addons/purchase/tests/__init__.py
+++ b/addons/purchase/tests/__init__.py
@@ -7,3 +7,4 @@ from . import test_purchase_invoice
 from . import test_access_rights
 from . import test_accrued_purchase_orders
 from . import test_purchase_tax_totals
+from . import test_purchase_dashboard