Add a new subtype_id for 'RFQ Sent' state. This removes
the need to join on mail.tracking.value when computing
'all_sent_rfqs'. Since mail.tracking.value is usually
a big table, removing this join leads to a substantial
speedup when loading the purchase.order dashboard on big
databases.

Add a unit test to check the dashboard values.

#### Speedup

The reported times were obtained using a v15 customer database with 4.5M mail_messages
and 4.75M mail_tracking_values. 2690 tracking_values were for RFQ Sent state. The data
was collected by timing queries in psql directly, so using a v15 DB is fine even if this PR targets master.

The reported times format is `cold cache time (hot cache time)`.

| Mail message and tracking | Before PR | After PR |
|:---------------------------:|:----------:|:---------:|
| 100 000 | 5s (15ms) | 1.1s (2ms) |
| 500 000 | 12s (20ms) | 1.3s (4ms) |
| 2M | 34s (30ms) | 1.6s (11ms) |
| 4.5M | 50s (45ms) | 2s (12ms) |


--
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr